# contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: MYJUNGLE Backend API
  version: 1.3.0
  description: |
    Contract for backoffice and MCP integrations.
    This document tracks the current `backend-api` router surface, including:
    session auth, platform management, SSH/API key management, project analysis,
    commit history, SSE streams, and both embedding and LLM provider settings.

    ## Error Status Codes
    - `400` - invalid JSON body or malformed query/path parameters
    - `401` - authentication required or invalid credentials
    - `403` - insufficient permissions
    - `404` - resource not found
    - `409` - conflict
    - `413` - request body too large
    - `422` - validation error
    - `500` - internal server error
    - `502` - upstream dependency unavailable
servers:
  - url: http://localhost:8080
    description: Local
security:
  - BearerAuth: []
  - CookieAuth: []
...
paths:
...
  /v1/platform-management/platform-roles:
    get:
      tags: [Platform Management]
      operationId: listPlatformRoles
      description: User session plus `platform_admin` role required.
      responses:
        "200":
          description: Active platform role assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformRoleListResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Platform Management]
      operationId: grantPlatformRole
      description: User session plus `platform_admin` role required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantPlatformRoleRequest"
      responses:
        "201":
          description: Platform role granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformRoleGrantResponse"
        "200":
          description: Platform role was already present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformRoleActionResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /v1/platform-management/platform-roles/{user_id}/{role}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/PlatformRole"
    delete:
      tags: [Platform Management]
      operationId: revokePlatformRole
      description: User session plus `platform_admin` role required.
      responses:
        "200":
          description: Platform role revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformRoleActionResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: Session token or API key via the `Authorization` header
    CookieAuth:
      type: apiKey
      in: cookie
      name: mj_session
      description: Session cookie set by `POST /v1/auth/login`
  parameters:
...
    UserID:
      in: path
      name: user_id
      required: true
      schema:
        type: string
        format: uuid
...
    PlatformRole:
      in: path
      name: role
      required: true
      schema:
        type: string
        enum: [platform_admin]
  responses:
    ErrorResponse:
      description: Structured error payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
...
    PlatformRoleActionResponse:
      type: object
      required: [message, user_id, role]
      properties:
        message:
          type: string
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [platform_admin]
...
    PlatformRoleAssignment:
      type: object
      required: [id, user_id, username, role, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
        role:
          type: string
          enum: [platform_admin]
        granted_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    PlatformRoleListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PlatformRoleAssignment"
    GrantPlatformRoleRequest:
      type: object
      required: [user_id, role]
      properties:
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [platform_admin]
    PlatformRoleGrantResponse:
      type: object
      required: [id, user_id, role, created_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [platform_admin]
        granted_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...